Add Code to Manually Check Services
Add JS/AJAX, Routes, Handlers, and DB funcs
to manually check an HTTP service. The result
of the check is updated in the database and
displayed back to the user on the page.
Changes in status are automatically reflected
as services are moved from one tab to
the appropriate tab (e.g., pending -> problem).